requirements.txt
Flask==0.10.1
Jinja2==2.8
speaklater==1.4
SQLAlchemy==1.0.11
WTForms==2.1
python_bcrypt==0.3.1
sqlalchemy_migrate==0.10.0
